Rename request package to exchange
