+++ exchange/build_test.go
@@ -1,4 +1,4 @@
-package request
+package exchange
 
 import (
 	"encoding/json"
