+++ exchange/options.go
@@ -1,4 +1,4 @@
-package request
+package exchange
 
 import "time"
